INtegrated the Ai to work with app.js
