+++ src/models/disaster.js
@@ -2,7 +2,7 @@ import mongoose from 'mongoose';
 
 const disasterSchema = new mongoose.Schema({
   name: { type: String, required: true },
-  type: { type: String, required: true },
+  type: { type: String, default: 'Disaster' },
   location: {
     type: {
       type: String, 
@@ -15,10 +15,9 @@ const disasterSchema = new mongoose.Schema({
     }
   },
   severity: { type: Number, required: true },
-  startDate: { type: Date, required: true },
-  endDate: { type: Date },
-  affectedAreas: [String],
-  description: String
+  startDate: { type: Date, default: Date.now },
+  description: String,
+  affectedAreas: [String]
 });
 
 disasterSchema.index({ location: '2dsphere' });
