--- test stuff/index.html
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Live Disaster Reports - Raipur</title>
-    <link rel="stylesheet" href="style.css">
-    <link rel="stylesheet" href="https://unpkg.com/leaflet/dist/leaflet.css" />
-</head>
-<body>
-    <h2>Live Disaster Reports - Raipur</h2>
-    <div class="container">
-        <div id="map"></div>
-        <div id="report">
-            <h3>Disaster Reports</h3>
-            <div id="report-list"></div>
-            <h3>Assistance Counters</h3>
-            <div id="locality-counters"></div>
-        </div>
-    </div>
-    <script src="https://unpkg.com/leaflet/dist/leaflet.js"></script>
-    <script src="https://cdn.socket.io/4.5.4/socket.io.min.js"></script>
-    <script src="map.js"></script>
-</body>
-</html>
